refactor(chat imports): update import paths for ChatRenderMessage
