+++ frontend/src/hooks/useChatSocket.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from 'react';
-import type { ChatServerMessage, ChatRenderMessage } from '@/types/chat';
+import type { ChatServerMessage, ChatRenderMessage } from '@/components/chat/types/chat';
 
 export function useChatSocket(
 	enabled: boolean,
